platzi-vision-api/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai==1.55.3
flask-cors==4.0.0
requests==2.31.0
